PR#9752 and PR#9699 (Apple silicon support from 4.12)
Revised handing of calling conventions for external C functions

(cherry picked from commit e41dc9c44327f40c1ab0feea12f779c931c9e31e)

Add support for iOS and macOS on ARM64

(cherry picked from commit a0a1ba4f1eb1a8a6c0013e77e3d6c1de5190c97f)

+++ runtime/memprof.c
@@ -130,7 +130,7 @@ static uintnat mt_generate_geom(void)
      bounded by the entropy provided by [mt_generate_uniform], which
      is 32bits. */
   res = 1 + logf(mt_generate_uniform()) * one_log1m_lambda;
-  if (res > Max_long) return Max_long;
+  if (res > (double)Max_long) return Max_long;
   return (uintnat)res;
 }
